Move Yard into documentation task
the yard gem is only loaded in the development environment.
In our `Rakefile` initializer we expect it to be there. On travis ci
only the
test environment is load. This cause the now the `Rakefile` to fail on
any task
as the `yard` can't be required:.
- update Rake to `12.2`
- remove old `-rubygems` parameter as of ruby `2.0` [this is not needed
anymore](http://guides.rubygems.org/make-your-own-gem/#your-first-gem)